sql/schemachanger/scplan/scviz: extract logic, lift embedded fields
This commit adds a new package which allows scgraphviz and sctestutils to share
code. The end result is that the embedded struct fields can be hoisted into
their parent for explain output.

Release note: None
