+++ pkg/BUILD.bazel
@@ -1516,6 +1516,7 @@ GO_TARGETS = [
     "//pkg/sql/schemachanger/scplan/internal/scgraph:scgraph_test",
     "//pkg/sql/schemachanger/scplan/internal/scgraphviz:scgraphviz",
     "//pkg/sql/schemachanger/scplan/internal/scstage:scstage",
+    "//pkg/sql/schemachanger/scplan/scviz:scviz",
     "//pkg/sql/schemachanger/scplan:scplan",
     "//pkg/sql/schemachanger/scplan:scplan_test",
     "//pkg/sql/schemachanger/screl:screl",
@@ -2494,6 +2495,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/sql/schemachanger/scplan/internal/scgraph:get_x_data",
     "//pkg/sql/schemachanger/scplan/internal/scgraphviz:get_x_data",
     "//pkg/sql/schemachanger/scplan/internal/scstage:get_x_data",
+    "//pkg/sql/schemachanger/scplan/scviz:get_x_data",
     "//pkg/sql/schemachanger/screl:get_x_data",
     "//pkg/sql/schemachanger/scrun:get_x_data",
     "//pkg/sql/schemachanger/sctest:get_x_data",
