+++ pkg/sql/schemachanger/scbuild/BUILD.bazel
@@ -72,7 +72,6 @@ go_test(
         "//pkg/server",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/parser",
-        "//pkg/sql/protoreflect",
         "//pkg/sql/schemachanger/rel",
         "//pkg/sql/schemachanger/scdeps/sctestdeps",
         "//pkg/sql/schemachanger/scdeps/sctestutils",
