+++ pkg/sql/schemachanger/scdeps/sctestutils/BUILD.bazel
@@ -15,17 +15,16 @@ go_library(
         "//pkg/sql/catalog/resolver",
         "//pkg/sql/faketreeeval",
         "//pkg/sql/parser",
-        "//pkg/sql/protoreflect",
         "//pkg/sql/schemachanger/scbuild",
         "//pkg/sql/schemachanger/scdeps",
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan",
+        "//pkg/sql/schemachanger/scplan/scviz",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
         "//pkg/sql/sessiondatapb",
         "//pkg/testutils/serverutils",
-        "//pkg/util/json",
         "//pkg/util/protoutil",
         "@com_github_kylelemons_godebug//diff",
         "@com_github_stretchr_testify//require",
