+++ pkg/sql/schemachanger/scplan/scviz/BUILD.bazel
@@ -1,0 +1,16 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "scviz",
+    srcs = ["maps.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scplan/scviz",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/util/protoutil",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_gogo_protobuf//jsonpb",
+    ],
+)
+
+get_x_data(name = "get_x_data")
