+++ pkg/sql/schemachanger/sctest/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//pkg/keys",
         "//pkg/sql/catalog",
         "//pkg/sql/parser",
-        "//pkg/sql/protoreflect",
         "//pkg/sql/schemachanger/scbuild",
         "//pkg/sql/schemachanger/scdecomp",
         "//pkg/sql/schemachanger/scdeps/sctestdeps",
@@ -26,6 +25,7 @@ go_library(
         "//pkg/sql/schemachanger/scop",
         "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scplan",
+        "//pkg/sql/schemachanger/scplan/scviz",
         "//pkg/sql/schemachanger/screl",
         "//pkg/sql/schemachanger/scrun",
         "//pkg/sql/sessiondata",
