+++ pkg/sql/schemachanger/testdata/explain_verbose/drop_index_hash_sharded_index
@@ -221,16 +221,15 @@ EXPLAIN (ddl, verbose) DROP INDEX idx CASCADE;
             ├── • LogEvent
             │     Element:
             │       SecondaryIndex:
-            │         embeddedIndex:
-            │           indexId: 2
-            │           isCreatedExplicitly: true
-            │           sharding:
-            │             columnNames:
-            │             - j
-            │             isSharded: true
-            │             name: crdb_internal_j_shard_16
-            │             shardBuckets: 16
-            │           tableId: 104
+            │         indexId: 2
+            │         isCreatedExplicitly: true
+            │         sharding:
+            │           columnNames:
+            │           - j
+            │           isSharded: true
+            │           name: crdb_internal_j_shard_16
+            │           shardBuckets: 16
+            │         tableId: 104
             │     EventBase:
             │       Authorization:
             │         UserName: root
