+++ pkg/sql/schemachanger/testdata/explain_verbose/drop_index_vanilla_index
@@ -131,10 +131,9 @@ EXPLAIN (ddl, verbose) DROP INDEX idx CASCADE;
             ├── • LogEvent
             │     Element:
             │       SecondaryIndex:
-            │         embeddedIndex:
-            │           indexId: 2
-            │           isCreatedExplicitly: true
-            │           tableId: 104
+            │         indexId: 2
+            │         isCreatedExplicitly: true
+            │         tableId: 104
             │     EventBase:
             │       Authorization:
             │         UserName: root
